Updated 2013, 2014 IP Charge and HRR files.
DRG 065 (intracranial hemorrhage) had a different description for 2011-2012 than it did for 2013-2014.  The DRG still covers the same condition, so the 2013 and 2014 IP charge data sets, as well as the HRR IP charge summary data set now have a DRG description consistent with the description in 2011-2012.  That is, all years (and HRR) have a single description for DRG 065: "065 - INTRACRANIAL HEMORRHAGE OR CEREBRAL INFARCTION W CC."